# libs/api-contract/src/main/resources/openapi/catalog-service-api.yaml
openapi: 3.0.3
info:
  title: Bookstore Catalog API
  version: 1.0.0
  description: >-
    REST API for managing the bookstore catalog. The catalog adopts Accept header based
    version negotiation (`application/vnd.vbookstore.catalog+json;version=1`) and uses
    optimistic concurrency via ETag headers on single resource representations.
...
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  headers:
    ETag:
      description: Entity tag for the returned representation.
      schema:
        type: string
  parameters:
    AcceptCatalog:
      name: Accept
      in: header
      required: true
      schema:
        type: string
        enum:
          - application/vnd.vbookstore.catalog+json;version=1
    ContentTypeCatalog:
      name: Content-Type
      in: header
      required: true
      schema:
        type: string
        enum:
          - application/vnd.vbookstore.catalog+json;version=1
    IfMatch:
      name: If-Match
      in: header
      required: true
      schema:
        type: string
    IfMatchOptional:
      name: If-Match
      in: header
      required: false
      schema:
        type: string
      description: Required when replacing an existing resource.
    IfNoneMatchStarOptional:
      name: If-None-Match
      in: header
      required: false
      schema:
        type: string
        enum:
          - '*'
      description: Send `*` to create a resource with a client supplied identifier.
    IfNoneMatchOptional:
      name: If-None-Match
      in: header
      required: false
      schema:
        type: string
    PageNumber:
      name: page[number]
      in: query
      description: Page index starting at 1.
      schema:
        type: integer
        minimum: 1
        default: 1
    PageSize:
      name: page[size]
      in: query
      description: Page size (max 100).
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
...
  schemas:
    Problem:
      type: object
      required:
        - title
        - status
      properties:
        type:
          type: string
          format: uri
        title:
          type: string
        status:
          type: integer
        detail:
          type: string
        instance:
          type: string
          format: uri
        code:
          type: string
        violations:
          type: array
          items:
            type: object
            properties:
              field:
                type: string
              message:
                type: string
...
    ResourceMetadata:
      type: object
      required:
        - createdAt
        - updatedAt
        - version
      properties:
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        version:
          type: integer
          format: int64
...
    PageMeta:
      type: object
      required:
        - totalElements
        - totalPages
        - page
        - size
      properties:
        totalElements:
          type: integer
          format: int64
        totalPages:
          type: integer
        page:
          type: integer
          description: Current page number (1-indexed).
        size:
          type: integer
...
    Author:
      type: object
      required:
        - id
        - name
        - metadata
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        metadata:
          $ref: '#/components/schemas/ResourceMetadata'
    AuthorRequest:
      type: object
      required:
        - name
      properties:
        name:
          type: string
    AuthorPatchRequest:
      type: object
      properties:
        name:
          type: string
    AuthorPage:
      type: object
      required:
        - content
        - meta
      properties:
        content:
          type: array
          items:
            $ref: '#/components/schemas/Author'
        meta:
          $ref: '#/components/schemas/PageMeta'
paths:
...
  /api/authors:
    get:
      tags: [Supporting Collections]
      summary: List authors
      operationId: listAuthors
      parameters:
        - $ref: '#/components/parameters/AcceptCatalog'
        - $ref: '#/components/parameters/PageNumber'
        - $ref: '#/components/parameters/PageSize'
      responses:
        '200':
          description: Paginated list of authors
          content:
            application/vnd.vbookstore.catalog+json;version=1:
              schema:
                $ref: '#/components/schemas/AuthorPage'
  /api/authors/{author_id}:
    parameters:
      - name: author_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
    put:
      tags: [Supporting Collections]
      summary: Create or replace an author
      operationId: putAuthor
      parameters:
        - $ref: '#/components/parameters/AcceptCatalog'
        - $ref: '#/components/parameters/ContentTypeCatalog'
        - $ref: '#/components/parameters/IfNoneMatchStarOptional'
        - $ref: '#/components/parameters/IfMatchOptional'
      requestBody:
        required: true
        content:
          application/vnd.vbookstore.catalog+json;version=1:
            schema:
              $ref: '#/components/schemas/AuthorRequest'
      responses:
        '200':
          description: Updated author representation
          headers:
            ETag:
              $ref: '#/components/headers/ETag'
          content:
            application/vnd.vbookstore.catalog+json;version=1:
              schema:
                $ref: '#/components/schemas/Author'
        '201':
          description: Author created
          headers:
            Location:
              description: Canonical URI of the created author.
              schema:
                type: string
            ETag:
              $ref: '#/components/headers/ETag'
          content:
            application/vnd.vbookstore.catalog+json;version=1:
              schema:
                $ref: '#/components/schemas/Author'
        '400':
          description: Validation error
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
        '404':
          description: Author not found
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
        '409':
          description: Author name conflict
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
        '412':
          description: Precondition failed
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
    get:
      tags: [Supporting Collections]
      summary: Retrieve author by identifier
      operationId: getAuthor
      parameters:
        - $ref: '#/components/parameters/AcceptCatalog'
        - $ref: '#/components/parameters/IfNoneMatchOptional'
      responses:
        '200':
          description: Author representation
          headers:
            ETag:
              $ref: '#/components/headers/ETag'
          content:
            application/vnd.vbookstore.catalog+json;version=1:
              schema:
                $ref: '#/components/schemas/Author'
        '304':
          description: Not modified
        '404':
          description: Author not found
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
    patch:
      tags: [Supporting Collections]
      summary: Update author metadata
      operationId: patchAuthor
      parameters:
        - $ref: '#/components/parameters/AcceptCatalog'
        - $ref: '#/components/parameters/ContentTypeCatalog'
        - $ref: '#/components/parameters/IfMatch'
      requestBody:
        required: true
        content:
          application/vnd.vbookstore.catalog+json;version=1:
            schema:
              $ref: '#/components/schemas/AuthorPatchRequest'
      responses:
        '200':
          description: Updated author representation
          headers:
            ETag:
              $ref: '#/components/headers/ETag'
          content:
            application/vnd.vbookstore.catalog+json;version=1:
              schema:
                $ref: '#/components/schemas/Author'
        '400':
          description: Validation error
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
        '404':
          description: Author not found
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
        '412':
          description: Precondition failed
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
    delete:
      tags: [Supporting Collections]
      summary: Delete author
      operationId: deleteAuthor
      parameters:
        - $ref: '#/components/parameters/AcceptCatalog'
        - $ref: '#/components/parameters/IfMatch'
      responses:
        '204':
          description: Author deleted
        '404':
          description: Author not found
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
        '412':
          description: Precondition failed
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'